1.4.4
1.用戶不能自己退出其所在的組
2.每個組添加改組存儲空間使用量

+++ apps/group_custom/ajax/delgroup.php
@@ -30,6 +30,13 @@ $l = OC_L10N::get('group_custom');
 
 if ( isset($_POST['group']) ) {
 
+	// Jawinton::begin
+	if ($_POST['group'] == 'admin') {
+		OCP\JSON::error(array('data' => array('title'=> $l->t('Delete Group') , 'message' => 'error' ))) ;
+		exit();
+	}
+	// Jawinton::end
+
     $result = OC_Group::deleteGroup( $_POST['group'] ) ;	// Jawtion
 
     if ($result) {
